Fix an implicit loss-of-precision warning in fraction
Fixes:

  warning: implicit conversion from 'const int64_t' (aka 'const
  long') to 'double' may lose precision

The fix is to make the cast explicit: yes, we're OK with possibly
losing 64-bit integer precision :-)

+++ include/fraction.h
@@ -77,12 +77,12 @@ public:
 
 	constexpr double ToDouble() const
 	{
-		return static_cast<double>(num) / denom;
+		return static_cast<double>(num) / static_cast<double>(denom);
 	}
 
 	constexpr float ToFloat() const
 	{
-		return static_cast<float>(num) / denom;
+		return static_cast<float>(num) / static_cast<float>(denom);
 	}
 
 	// Equality
